% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGOTable.R
\name{createGOTable}
\alias{createGOTable}
\title{createGOTable}
\usage{
createGOTable(
  all_genes,
  shown_terms = 20,
  min_signficant = 5,
  select_ontology = "BP",
  select_organism = "org.Mm.eg.db",
  select_ID = "SYMBOL"
)
}
\description{
creates a Gene Ontology enrichment table on DESeq2 results using topGO package
}
