% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{getResults}
\usage{
getResults(
  dds,
  contrast,
  lfc_cutoff = 0,
  shrink = FALSE,
  annotation = "gtf",
  anno_id = "gene_id",
  anno_symbol = "gene_symbol"
)
}
\description{
get results for comparisons
}
