% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGOTable.R
\name{makeGOTable}
\alias{makeGOTable}
\title{makeGOTable}
\usage{
makeGOTable(
  all_genes,
  shown_terms = 20,
  min_signficant = 5,
  select_ontology = "BP",
  select_organism = "org.Mm.eg.db",
  select_ID = "SYMBOL"
)
}
\description{
make a Gene Ontology enrichment table on DESeq2 results using topGO package
}
