% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDens.R
\name{plotDens}
\alias{plotDens}
\title{plotDens}
\usage{
plotDens(
  res,
  main_title = "",
  selection_ids = NULL,
  selection_id_type = "symbol",
  selection_color = rgb(0.7, 0, 0.9, 1),
  selection_legend = NULL,
  xlims = c(-5, 5),
  ylims = c(0, 1),
  x_label = "log2FC"
)
}
\description{
plot density of log2FC values from results table
}
