% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOGraph.R
\name{plotGOGraph}
\alias{plotGOGraph}
\title{plotGOGraph}
\usage{
plotGOGraph(
  all_genes,
  first_sig_nodes = 5,
  select_ontology = "BP",
  select_organism = "org.Mm.eg.db",
  select_ID = "SYMBOL"
)
}
\description{
make a Gene Ontology enrichment table on DESeq2 results using topGO package
}
