% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadStats.R
\name{plotReadStats}
\alias{plotReadStats}
\title{plotReadStats}
\usage{
plotReadStats(
  read.counts,
  conditions,
  bar_colors = c("#999999", "#0072B2", "#CC79A7", "#009E73", "#E69F00", "#D55E00",
    "#56B4E9", "#F0E442")
)
}
\description{
get read stats from count table generated by STAR aligner
}
