% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMA.R
\name{plottingMA}
\alias{plottingMA}
\title{plottingMA}
\usage{
plottingMA(
  res,
  main_title = "",
  main_title_size = 1,
  point_size = 0.25,
  point_color = rgb(0.7, 0.7, 0.7, 0.5),
  sign_point_color = rgb(0.9, 0.6, 0, 0.5),
  selection_ids = NULL,
  selection_name = "labeled",
  selection_id_type = "symbol",
  selection_point_size = 0.5,
  selection_point_color = rgb(0, 0, 0, 1),
  selection_sign_point_color = rgb(0.8, 0, 0, 1),
  selection_text_label = FALSE,
  selection_text_size = 1,
  selection_shadow = FALSE,
  xlims = c(0, 6),
  ylims = c(-5, 5),
  x_axis_by = 2,
  padj_cutoff = 0.01,
  show_legend = TRUE
)
}
\description{
compare log2FC and baseMean values from results table
}
