% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMAold.R
\name{plottingMAold}
\alias{plottingMAold}
\title{plottingMAold}
\usage{
plottingMAold(
  res,
  main_title = "",
  selection_ids = NULL,
  selection_id_type = "symbol",
  selection_point_size = 0.5,
  selection_text_label = FALSE,
  selection_shadow = FALSE,
  xlims = c(0, 6),
  ylims = c(-5, 5),
  x_axis_by = 2,
  padj_cutoff = 0.01,
  show_legend = TRUE
)
}
\description{
compare log2FC and baseMean values from results table
}
