% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDDS.R
\name{setupDDS}
\alias{setupDDS}
\title{setupDDS}
\usage{
setupDDS(
  SampleTableName,
  CountTableName,
  SampleIdName,
  ConditionName,
  BatchName = NULL,
  n_samples_for_filtering = 3,
  min_number_of_reads = 1
)
}
\description{
setup dds object
}
