% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDDSwith3wayInteraction.R
\name{setupDDSwith3wayInteraction}
\alias{setupDDSwith3wayInteraction}
\title{setupDDSwith3wayInteraction}
\usage{
setupDDSwith3wayInteraction(
  SampleTableName,
  CountTableName,
  SampleIdName,
  FactorAName,
  FactorBName,
  FactorCName,
  BatchName = NULL,
  n_samples_for_filtering = 3,
  min_number_of_reads = 1
)
}
\description{
setup dds object with 3-way interaction term
}
