% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDDSwithInteraction.R
\name{setupDDSwithInteraction}
\alias{setupDDSwithInteraction}
\title{setupDDSwithInteraction}
\usage{
setupDDSwithInteraction(
  SampleTableName,
  CountTableName,
  SampleIdName,
  FactorAName,
  FactorBName,
  BatchName = NULL,
  n_samples_for_filtering = 3,
  min_number_of_reads = 1
)
}
\description{
setup dds object with interaction term
}
