% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLog2FC.R
\name{plotLog2FC}
\alias{plotLog2FC}
\title{plotLog2FC}
\usage{
plotLog2FC(
  res1,
  res2,
  main_title = "",
  x_label = "log2FC",
  y_label = "log2FC",
  lims = c(-5, 5),
  point_size = 0.25,
  selection_ids = NULL,
  selection_id_type = "symbol",
  selection_color = rgb(0.7, 0, 0.9, 1),
  selection_point_size = 0.5,
  selection_legend = NULL,
  selection_text_label = FALSE
)
}
\description{
compare log2FC values from two results table
}
